requirements.txt
django
gunicorn
django-heroku
django-cors-headers
django-rest-framework
PyJWT









